Issue #5303 move existing ManagedCompletableFuture impl to the concurrent.mp bundle
