+++ dev/com.ibm.ws.concurrent.mp.1.0/bnd.bnd
@@ -23,13 +23,19 @@ Require-Capability: osgi.ee;filter:="(&(osgi.ee=JavaSE)(version=1.8))"
 WS-TraceGroup: concurrent
 
 Export-Package:\
-  com.ibm.ws.concurrent.mp
+  com.ibm.ws.concurrent.mp,\
+  com.ibm.ws.concurrent.rx
+  # TODO remove the above export once ManagedCompletableFuture is fully transitioned to MicroProfile Concurrency interfaces
     
 instrument.classesExcludes: com/ibm/ws/concurrent/mp/resources/*.class
 
 -buildpath: \
   com.ibm.websphere.appserver.spi.logging,\
+  com.ibm.websphere.javaee.concurrent.1.0;version=latest,\
   com.ibm.websphere.org.eclipse.microprofile.concurrency.1.0,\
+  com.ibm.websphere.org.osgi.core;version=latest,\
   com.ibm.ws.concurrent;version=latest,\
   com.ibm.ws.context;version=latest,\
-  com.ibm.ws.org.osgi.annotation.versioning;version=latest
+  com.ibm.ws.kernel.service;version=latest,\
+  com.ibm.ws.org.osgi.annotation.versioning;version=latest,\
+  com.ibm.ws.threading;version=latest
