+++ dev/com.ibm.ws.concurrent_fat_rx/publish/files/features/managedCompletableFuture-1.0.mf
@@ -1,7 +1,7 @@
 Subsystem-ManifestVersion: 1
 Subsystem-SymbolicName: managedCompletableFuture-1.0; visibility:=public
 Subsystem-Version: 1.0
-Subsystem-Content: com.ibm.ws.concurrent.rx
+Subsystem-Content: com.ibm.ws.concurrent.mp.1.0
 Subsystem-Type: osgi.subsystem.feature
 IBM-Feature-Version: 2
-IBM-API-Package: com.ibm.websphere.concurrent.rx; type="ibm-api"
+IBM-API-Package: com.ibm.ws.concurrent.rx; type="ibm-api"
